+++ src/main/java/buaa/bp/asclepius/mapper/AppointmentMapper.java
@@ -5,9 +5,9 @@ import java.util.List;
 import buaa.bp.asclepius.model.Appointment;
 
 public interface AppointmentMapper {
-	public List<Appointment> getAllAppointments(int userId);
-	public Appointment getAppointmentById(int userId,int appointmentId);
+	public List<Appointment> getAllAppointments(long userId);
+	public Appointment getAppointmentById(long userId,long appointmentId);
 	public int updateAppointment(Appointment appointment);
 	public int createAppointment(Appointment appointment);
-	public int deleteAppoinement(int id);
+	public int deleteAppoinement(long id);
 }
