+++ src/main/java/buaa/bp/asclepius/mapper/UserMapper.java
@@ -6,9 +6,9 @@ import buaa.bp.asclepius.model.User;
 
 public interface UserMapper {
 	public List<User> getAllUsers();
-	public User getUserById(int id);
+	public User getUserById(long id);
 	public int createUser(User user);
 	public int updateUser(User user);
-	public int deleteUser(int id);
+	public int deleteUser(long id);
 	public User getUserByName(String name);
 }
